IMPALA-7802: Close connections of idle client sessions
Previously, if idle session timeout is set either via
startup flag or query options, a client session will expire
after that set period of inactivity. However, the network
connection and the service thread of an expired session will
still be around until the session is closed by the client.
This is highly undesirable as these idle sessions still count
towards the quota bound by --fe_esrvice_threads, so if the
total number of sessions (including the idle ones) reaches
that upper bound, all incoming new session will block until
some of the existing sessions exit. There is no time bound on
when those expired sessions will be closed. In some sense,
leaving many idle sessions opened is a denial-of-service attack
on Impala.

This change implements support for closing expired client sessions.
In particular, a new flag --idle_client_poll_time_s is added to
specify a time interval in seconds of client's inactivity which
will cause an idle service thread of a client connection to wake up
and check if all sessions associated with the connection are idle.
If so, the connection will be closed. This allows the service threads
to be freed up without waiting for client to close the connections.

Testing done:
- core build
- new targeted test which verifies the connections of expired sessions
are closed.
- verified the flags function as expected in a secure cluster with Kerberos + SSL

Cherry-pick notes:
1. Cherry-pick was not clean since the constructor signatures did not match with the original
patch in TAcceptQueueServer. The conflict was resolved by modifying the revelant
constructor to pass the parameter for idle_client_poll_period
2. There were conflicts in the test since the C5 test framework does not have the HS2
client provided from the impala_test_suite.py. The logic to open a session and run a query
on HS2 port was added separately in a private method in the test.
3. Test do not accept --cluster_size arg in CDH5. This was worked-around by choosing the
first impalad in the test instead of a random one.

Change-Id: I0e15f603ef538990708c5703131b656ed46905b4
Reviewed-on: http://gerrit.cloudera.org:8080/13607
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 3d9a6763afb54cf0e79b559ef87c42a1724e4e11)