IMPALA-8635. Use local metastore URIs when checking for Kudu/HMS integration
The patch for IMPALA-8504 (part 2) (6bb404dc359) checks to see if Impala
and Kudu are configured against the same metastore to determine if the
HMS integration is enabled. However, instead of using its own metastore
URI config, it uses the configuration stored on the remote HMS. This is
error prone because it's not common for the HMS configuration to store
its own URI. Instead, we should use our own config.

This patch changes to using the local configuration for this purpose.
More robust would be to use the HMS "UUID" support, since it's possible
that Kudu and Impala are talking to different HMS instances sharing a
backing DB, but that work is deferred to a later commit since it depends
on Kudu-side changes.

This commit doesn't add any tests, but fixes the existing tests when
running against Hive 3, where the HMS server side uses a different
configuration key for the metastore URIs.

Change-Id: Id7a4c2cc0580f7c4dc5cfceed30b91e87c547612
Reviewed-on: http://gerrit.cloudera.org:8080/13555
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Thomas Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/48525
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Hao Hao <hao.hao@cloudera.com>