CDH-80678: Fix INTERNAL_LISTEN_(HOST|PORT) for test_kudu_alter_table
TestKuduHMSIntegration.test_kudu_alter_table was added by IMPALA-8506,
which was backported to 6.3.x. The test relies on replacing
INTERNAL_LISTEN_(HOST|PORT) in test files, which was added in
IMPALA-7995, which was not backported to 6.3.x. As a result, the test
fails.

This patch backports the relevant parts of MPALA-7995.

Change-Id: Ia46dd91f971c05be5cf425bc935e71421a95049b
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49110
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: quasar-precommit User
Reviewed-by: Andrew Sherman <asherman@cloudera.com>